# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paqar Public API",
    "version": "1.0.0",
    "description": "Malaysian used-car valuation and variant guide API for LLMs and integrations",
    "contact": {
      "email": "support@paqar.my"
    }
  },
  "servers": [
    {
      "url": "https://paqar.my/api/v1",
      "description": "Production API"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development"
    }
  ],
  "paths": {
    "/plate/{plate}": {
      "get": {
        "summary": "Get vehicle teaser by plate",
        "tags": ["Plate Lookup"],
        "parameters": [
          {
            "name": "plate",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "WPH925",
            "description": "Vehicle plate (3 letters + 3 digits)"
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "make": { "type": "string" },
                    "model": { "type": "string" },
                    "registrationYear": { "type": "string" },
                    "color": { "type": "string" },
                    "mileage": { "type": "number" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid plate format" },
          "404": { "description": "Plate not found" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/valuation": {
      "get": {
        "summary": "Get full valuation",
        "tags": ["Valuation"],
        "parameters": [
          {
            "name": "plate",
            "in": "query",
            "schema": { "type": "string" },
            "example": "WPH925"
          },
          {
            "name": "nvic",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "make",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "year",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "model",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Valuation found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "variant": { "type": ["string", "null"] },
                    "wmNewPrice": { "type": ["number", "null"] },
                    "marketMedian": { "type": ["number", "null"] },
                    "marketMin": { "type": ["number", "null"] },
                    "marketMax": { "type": ["number", "null"] },
                    "marketCount": { "type": "number" },
                    "confidence": { "type": "string", "enum": ["high", "medium", "low"], "description": "Weight of the comparable set. high = 10+ listings, medium = 5-9, low = 0-4. Mixed-variant cohorts are capped at medium. Separate from whether Paqar would issue a buyer-facing verdict." },
                    "isSpecialVariant": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing parameters" },
          "404": { "description": "Vehicle or valuation not found" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/variants/{make}/{model}": {
      "get": {
        "summary": "Get variant guide",
        "tags": ["Variants"],
        "parameters": [
          {
            "name": "make",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "Honda"
          },
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "City"
          }
        ],
        "responses": {
          "200": {
            "description": "Variant guide found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "model": { "type": "string" },
                    "modelSlug": { "type": "string" },
                    "generations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "years": { "type": "string" },
                          "variants": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": { "type": "string" },
                                "verdict": { "type": "string" },
                                "spotChecks": { "type": "array", "items": { "type": "string" } }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Model not found" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    }
  },
  "components": {
    "headers": {
      "X-Citation": {
        "schema": { "type": "string" },
        "description": "Attribution header for API responses"
      },
      "Retry-After": {
        "schema": { "type": "integer" },
        "description": "Seconds until rate limit resets"
      }
    }
  }
}